requirements-github-actions.txt
PyGithub
requests
numpy
pandas
